Working on type checking, printf

+++ src/core/irvariabledescriptor.h
@@ -18,7 +18,7 @@ public:
     std::string Get()
     {
         std::string builder = "";
-        builder += IsPointer ? "ptr" : "i32";
+        builder += IsPointer ? "ptr" : "i64";
         builder += " ";
         builder += IsConstant ? "@" : "%";
         builder += Name;
